add tests to Event and Parser classes
