+++ src/main/java/duke/Deadline.java
@@ -1,0 +1,33 @@
+package duke;
+
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
+
+public class Deadline extends Task {
+
+    Deadline(String description, String by, boolean isDone) {
+        super(description, isDone);
+        String date;
+        try {
+            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+            date = LocalDate.parse(by, formatter).format(DateTimeFormatter.ofPattern("MMM dd yyyy"));
+        } catch (DateTimeParseException e) {
+            date = by;
+        }
+
+        this.by = date;
+    }
+
+    protected String by;
+
+    public Deadline(String description, String by) {
+        this(description, by, false);
+
+    }
+
+    @Override
+    public String toString() {
+        return "[D]" + super.toString() + " (by: " + by + ")";
+    }
+}
