+++ src/main/java/duke/DukeException.java
@@ -1,0 +1,7 @@
+package duke;
+
+public class DukeException extends RuntimeException {
+    public DukeException(String errorMessage) {
+        super(errorMessage);
+    }
+}
