+++ src/main/java/duke/Event.java
@@ -1,0 +1,32 @@
+package duke;
+
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
+
+public class Event extends Task {
+
+    protected String at;
+
+    public Event(String description, String at, boolean isDone) {
+        super(description, isDone);
+        String date;
+        try {
+            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+            date = LocalDate.parse(at, formatter).format(DateTimeFormatter.ofPattern("MMM dd yyyy"));
+        } catch (DateTimeParseException e) {
+            date = at;
+        }
+
+        this.at = date;
+    }
+    public Event(String description, String at) {
+        this(description, at, false);
+    }
+
+
+    @Override
+    public String toString() {
+        return "[E]" + super.toString() + " (at: " + at + ")";
+    }
+}
