+++ src/main/java/duke/Todo.java
@@ -1,0 +1,19 @@
+package duke;
+
+public class Todo extends Task {
+
+    protected String by;
+
+    public Todo(String description, boolean isDone) {
+        super(description, isDone);
+    }
+
+    public Todo(String description) {
+        this(description, false);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
