+++ src/main/java/duke/Ui.java
@@ -1,0 +1,35 @@
+package duke;
+
+public class Ui {
+
+    private static void printLine() {
+        for (int i = 0; i < 20; i++) {
+            System.out.print("-");
+        }
+        System.out.println();
+    }
+
+    public static void initialize() {
+        String logo = " ____        _        \n"
+                + "|  _ \\ _   _| | _____ \n"
+                + "| | | | | | | |/ / _ \\\n"
+                + "| |_| | |_| |   <  __/\n"
+                + "|____/ \\__,_|_|\\_\\___|\n";
+        System.out.println("Hello from\n" + logo);
+        printStatement("Hello! I'm Duke\nWhat can I do for you?\n");
+    }
+
+    public static void printStatement(String statement) {
+        System.out.println();
+        printLine();
+//        printPadding();
+        System.out.println(statement);
+        System.out.println();
+        printLine();
+        System.out.println();
+    }
+
+    public static void printError(String errorMsg) {
+        printStatement("OOPS! " + errorMsg);
+    }
+}
