+++ src/test/java/duke/DukeTest.java
@@ -1,0 +1,20 @@
+package duke;
+import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
+
+
+public class DukeTest {
+    @Test
+    public void dummyTest() {
+        assertEquals(2, 2);
+    }
+
+    @Test
+    public void startUpTest() {
+        Duke duke = new Duke();
+        duke.run();
+    }
+
+
+}
